+++ docs/app/documentation/component-docs/side-navigation/examples/side-navigation-examples.component.ts
@@ -1,20 +1,33 @@
-import { Component } from '@angular/core';
+import { Component, ViewEncapsulation } from '@angular/core';
 
 @Component({
     selector: 'fd-side-navigation-example',
-    templateUrl: './side-navigation-example.component.html'
+    templateUrl: './side-navigation-example.component.html',
 })
 export class SideNavigationExampleComponent {}
 
 @Component({
     selector: 'fd-side-navigation-collapsed-example',
-    templateUrl: './side-navigation-collapsed-example.component.html'
+    templateUrl: './side-navigation-collapsed-example.component.html',
+    styles: [`
+        .fd-side-nav--icons {
+            max-width: 65px;
+        }
+    `],
+    encapsulation: ViewEncapsulation.None
 })
 export class SideNavigationCollapsedExampleComponent {}
 
 @Component({
     selector: 'fd-side-navigation-icons-example',
-    templateUrl: './side-navigation-icons-example.component.html'
+    templateUrl: './side-navigation-icons-example.component.html',
+    styles: [
+        `
+            fd-icon {
+                padding-right: 10px;
+            }
+        `
+    ]
 })
 export class SideNavigationIconsExampleComponent {}