+++ docs/app/documentation/core-helpers/component-example/component-example.component.ts
@@ -7,7 +7,7 @@ declare let hljs: any;
     selector: 'component-example',
     template: `
         <div class="fd-tile docs-component docs-component__ fd-has-background-color-background-1" id="{{id}}">
-            <div class="fd-tile__content">
+            <div class="fd-tile__content fd-tile-content-example">
                 <div class="component-example__features">
                     <rtl-toggle [label]="id2"></rtl-toggle>
                     <background-toggle [label]="id"></background-toggle>
@@ -27,6 +27,10 @@ declare let hljs: any;
 
             rtl-toggle {
                 padding-right: 20px;
+            }
+
+            .fd-tile-content-example {
+                padding: 20px;
             }
         `
     ]