+++ docs/app/documentation/documentation.routes.ts
@@ -11,7 +11,7 @@ import { BreadcrumbDocsComponent } from './component-docs/breadcrumb/breadcrumb-
 import { ButtonDocsComponent } from './component-docs/button/button-docs.component';
 import { ButtonGroupDocsComponent } from './component-docs/button-group/button-group-docs.component';
 import { CalendarDocsComponent } from './component-docs/calendar/calendar-docs.component';
-import { ComboboxInputDocsComponent } from './component-docs/combobox-input/combobox-input-docs.component';
+import { ComboboxDocsComponent } from './component-docs/combobox/combobox-docs.component';
 import { DatePickerDocsComponent } from './component-docs/date-picker/date-picker-docs.component';
 import { DatetimePickerDocsComponent } from './component-docs/datetime-picker/datetime-picker-docs.component';
 import { DropdownDocsComponent } from './component-docs/dropdown/dropdown-docs.component';
@@ -51,7 +51,7 @@ import { BreadcrumbHeaderComponent } from './component-docs/breadcrumb/breadcrum
 import { ButtonHeaderComponent } from './component-docs/button/button-header/button-header.component';
 import { ButtonGroupHeaderComponent } from './component-docs/button-group/button-group-header/button-group-header.component';
 import { CalendarHeaderComponent } from './component-docs/calendar/calendar-header/calendar-header.component';
-import { ComboboxInputHeaderComponent } from './component-docs/combobox-input/combobox-input-header/combobox-input-header.component';
+import { ComboboxHeaderComponent } from './component-docs/combobox/combobox-header/combobox-header.component';
 import { DatePickerHeaderComponent } from './component-docs/date-picker/date-picker-header/date-picker-header.component';
 import { DatetimePickerHeaderComponent } from './component-docs/datetime-picker/datetime-picker-header/datetime-picker-header.component';
 import { DropdownHeaderComponent } from './component-docs/dropdown/dropdown-header/dropdown-header.component';
@@ -84,6 +84,12 @@ import { TimePickerHeaderComponent } from './component-docs/time-picker/time-pic
 import { TreeHeaderComponent } from './component-docs/tree/tree-header/tree-header.component';
 import { ToggleHeaderComponent } from './component-docs/toggle/toggle-header/toggle-header.component';
 import { TokenHeaderComponent } from './component-docs/token/token-header/token-header.component';
+import { SplitButtonHeaderComponent } from './component-docs/split-button/split-button-header/split-button-header.component';
+import { SplitButtonDocsComponent } from './component-docs/split-button/split-button-docs.component';
+import {
+    TabNavigationExampleChildFirst,
+    TabNavigationExampleChildSecond, TabNavigationExampleChildThird
+} from './component-docs/tabs/examples/tab-navigation-children/tab-navigation-children';
 import { SelectHeaderComponent } from './component-docs/select/select-header/select-header.component';
 import { SelectDocsComponent } from './component-docs/select/select-docs.component';
 
@@ -93,250 +99,215 @@ export const ROUTES: Routes = [
         component: DocumentationComponent,
         children: [
             { path: '', redirectTo: 'home', pathMatch: 'full' },
-            {
-                path: 'action-bar', component: ActionBarHeaderComponent, children: [
-                    { path: '', component: ActionBarDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.actionBar } }
+            { path: 'action-bar', component: ActionBarHeaderComponent, children: [
+                    { path: '', component: ActionBarDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.actionBar}}
                 ]
             },
-            {
-                path: 'alert', component: AlertHeaderComponent, children: [
-                    { path: '', component: AlertDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.alert } }
+            { path: 'alert', component: AlertHeaderComponent, children: [
+                    { path: '', component: AlertDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.alert}}
                 ]
             },
-            {
-                path: 'badgeLabel', component: BadgeLabelHeaderComponent, children: [
-                    { path: '', component: BadgeLabelDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.badgeLabel } }
+            { path: 'badgeLabel', component: BadgeLabelHeaderComponent, children: [
+                    { path: '', component: BadgeLabelDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.badgeLabel}}
                 ]
             },
-            {
-                path: 'breadcrumb', component: BreadcrumbHeaderComponent, children: [
-                    { path: '', component: BreadcrumbDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.breadcrumb } }
+            { path: 'breadcrumb', component: BreadcrumbHeaderComponent, children: [
+                    { path: '', component: BreadcrumbDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.breadcrumb}}
                 ]
             },
-            {
-                path: 'button', component: ButtonHeaderComponent, children: [
-                    { path: '', component: ButtonDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.button } }
+            { path: 'button', component: ButtonHeaderComponent, children: [
+                    { path: '', component: ButtonDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.button}}
                 ]
             },
-            {
-                path: 'buttonGroup', component: ButtonGroupHeaderComponent, children: [
-                    { path: '', component: ButtonGroupDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.buttonGroup } }
+            { path: 'buttonGroup', component: ButtonGroupHeaderComponent, children: [
+                    { path: '', component: ButtonGroupDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.buttonGroup}}
                 ]
             },
-            {
-                path: 'calendar', component: CalendarHeaderComponent, children: [
-                    { path: '', component: CalendarDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.calendar } }
+            { path: 'calendar', component: CalendarHeaderComponent, children: [
+                    { path: '', component: CalendarDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.calendar}}
                 ]
             },
-            {
-                path: 'comboboxInput', component: ComboboxInputHeaderComponent, children: [
-                    { path: '', component: ComboboxInputDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.comboboxInput } }
+            { path: 'combobox', component: ComboboxHeaderComponent, children: [
+                    { path: '', component: ComboboxDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.comboboxInput}}
                 ]
             },
-            {
-                path: 'datePicker', component: DatePickerHeaderComponent, children: [
-                    { path: '', component: DatePickerDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.datePicker } }
+            { path: 'datePicker', component: DatePickerHeaderComponent, children: [
+                    { path: '', component: DatePickerDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.datePicker}}
                 ]
             },
-            {
-                path: 'datetime-picker', component: DatetimePickerHeaderComponent, children: [
-                    { path: '', component: DatetimePickerDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.datetimePicker } }
+            { path: 'datetime-picker', component: DatetimePickerHeaderComponent, children: [
+                    { path: '', component: DatetimePickerDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.datetimePicker}}
                 ]
             },
-            {
-                path: 'dropdown', component: DropdownHeaderComponent, children: [
-                    { path: '', component: DropdownDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: [] } }
+            { path: 'dropdown', component: DropdownHeaderComponent, children: [
+                    { path: '', component: DropdownDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: []}}
                 ]
             },
-            {
-                path: 'file-input', component: FileInputHeaderComponent, children: [
-                    { path: '', component: FileInputDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.fileInput } }
+            { path: 'file-input', component: FileInputHeaderComponent, children: [
+                    { path: '', component: FileInputDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.fileInput}}
                 ]
             },
-            {
-                path: 'form', component: FormHeaderComponent, children: [
-                    { path: '', component: FormDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.form } }
+            { path: 'form', component: FormHeaderComponent, children: [
+                    { path: '', component: FormDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.form}}
                 ]
             },
-            {
-                path: 'icon', component: IconHeaderComponent, children: [
-                    { path: '', component: IconDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.icon } }
+            { path: 'icon', component: IconHeaderComponent, children: [
+                    { path: '', component: IconDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.icon}}
                 ]
             },
-            {
-                path: 'identifier', component: IdentifierHeaderComponent, children: [
-                    { path: '', component: IdentifierDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.identifier } }
+            { path: 'identifier', component: IdentifierHeaderComponent, children: [
+                    { path: '', component: IdentifierDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.identifier}}
                 ]
             },
-            {
-                path: 'image', component: ImageHeaderComponent, children: [
-                    { path: '', component: ImageDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.image } }
+            { path: 'image', component: ImageHeaderComponent, children: [
+                    { path: '', component: ImageDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.image}}
                 ]
             },
-            {
-                path: 'infiniteScroll', component: InfiniteScrollHeaderComponent, children: [
-                    { path: '', component: InfiniteScrollDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.infiniteScroll } }
+            { path: 'infiniteScroll', component: InfiniteScrollHeaderComponent, children: [
+                    { path: '', component: InfiniteScrollDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.infiniteScroll}}
                 ]
             },
-            {
-                path: 'inlineHelp', component: InlineHelpHeaderComponent, children: [
-                    { path: '', component: InlineHelpDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.inlineHelp } }
+            { path: 'inlineHelp', component: InlineHelpHeaderComponent, children: [
+                    { path: '', component: InlineHelpDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.inlineHelp}}
                 ]
             },
-            {
-                path: 'inputGroup', component: InputGroupHeaderComponent, children: [
-                    { path: '', component: InputGroupDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.inputGroup } }
+            { path: 'inputGroup', component: InputGroupHeaderComponent, children: [
+                    { path: '', component: InputGroupDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.inputGroup}}
                 ]
             },
-            {
-                path: 'list', component: ListHeaderComponent, children: [
-                    { path: '', component: ListDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.list } }
+            { path: 'list', component: ListHeaderComponent, children: [
+                    { path: '', component: ListDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.list}}
                 ]
             },
-            {
-                path: 'loadingSpinner', component: LoadingSpinnerHeaderComponent, children: [
-                    { path: '', component: LoadingSpinnerDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.loadingSpinner } }
+            { path: 'loadingSpinner', component: LoadingSpinnerHeaderComponent, children: [
+                    { path: '', component: LoadingSpinnerDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.loadingSpinner}}
                 ]
             },
-            {
-                path: 'menu', component: MenuHeaderComponent, children: [
-                    { path: '', component: MenuDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.menu } }
+            { path: 'menu', component: MenuHeaderComponent, children: [
+                    { path: '', component: MenuDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.menu}}
                 ]
             },
-            {
-                path: 'modal', component: ModalDocsHeaderComponent, children: [
-                    { path: '', component: ModalDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.modal } }
+            { path: 'modal', component: ModalDocsHeaderComponent, children: [
+                    { path: '', component: ModalDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.modal}}
                 ]
             },
-            {
-                path: 'multi-input', component: MultiInputHeaderComponent, children: [
-                    { path: '', component: MultiInputDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.multiInput } }
+            { path: 'multi-input', component: MultiInputHeaderComponent, children: [
+                    { path: '', component: MultiInputDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.multiInput}}
                 ]
             },
-            {
-                path: 'pagination', component: PaginationHeaderComponent, children: [
-                    { path: '', component: PaginationDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.pagination } }
+            { path: 'pagination', component: PaginationHeaderComponent, children: [
+                    { path: '', component: PaginationDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.pagination}}
                 ]
             },
-            {
-                path: 'panel', component: PanelDocsHeaderComponent, children: [
-                    { path: '', component: PanelDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.panel } }
+            { path: 'panel', component: PanelDocsHeaderComponent, children: [
+                    { path: '', component: PanelDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.panel}}
                 ]
             },
-            {
-                path: 'popover', component: PopoverHeaderComponent, children: [
-                    { path: '', component: PopoverDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.popover } }
+            { path: 'popover', component: PopoverHeaderComponent, children: [
+                    { path: '', component: PopoverDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.popover}}
                 ]
             },
-            {
-                path: 'popover-directive', component: PopoverDirectiveHeaderComponent, children: [
-                    { path: '', component: PopoverDirectiveDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.popoverDirective } }
+            { path: 'popover-directive', component: PopoverDirectiveHeaderComponent, children: [
+                    { path: '', component: PopoverDirectiveDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.popoverDirective}}
                 ]
             },
-            {
-                path: 'scroll-spy', component: ScrollSpyHeaderComponent, children: [
-                    { path: '', component: ScrollSpyDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.scrollSpy } }
+            { path: 'scroll-spy', component: ScrollSpyHeaderComponent, children: [
+                    { path: '', component: ScrollSpyDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.scrollSpy}}
                 ]
             },
-            {
-                path: 'searchInput', component: SearchInputHeaderComponent, children: [
-                    { path: '', component: SearchInputDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.searchInput } }
+            { path: 'searchInput', component: SearchInputHeaderComponent, children: [
+                    { path: '', component: SearchInputDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.searchInput}}
                 ]
             },
-            {
-                path: 'select', component: SelectHeaderComponent, children: [
-                    { path: '', component: SelectDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.select } }
+            { path: 'shellbar', component: ShellbarDocsHeaderComponent, children: [
+                    { path: '', component: ShellbarDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.shellbar}}
                 ]
             },
-            {
-                path: 'shellbar', component: ShellbarDocsHeaderComponent, children: [
-                    { path: '', component: ShellbarDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.shellbar } }
+            { path: 'sideNavigation', component: SideNavigationHeaderComponent, children: [
+                    { path: '', component: SideNavigationDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.sideNavigation}}
                 ]
             },
-            {
-                path: 'sideNavigation', component: SideNavigationHeaderComponent, children: [
-                    { path: '', component: SideNavigationDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.sideNavigation } }
+            { path: 'splitButton', component: SplitButtonHeaderComponent, children: [
+                    { path: '', component: SplitButtonDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.splitButton}}
                 ]
             },
-            {
-                path: 'table', component: TableDocsHeaderComponent, children: [
-                    { path: '', component: TableDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.table } }
+            { path: 'table', component: TableDocsHeaderComponent, children: [
+                    { path: '', component: TableDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.table}}
                 ]
             },
-            {
-                path: 'tabs', component: TabsHeaderComponent, children: [
-                    { path: '', component: TabsDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.tabs } }
+            { path: 'tabs', component: TabsHeaderComponent, children: [
+                    { path: '', component: TabsDocsComponent,
+                        children: [
+                            {path: 'tab1', component: TabNavigationExampleChildFirst},
+                            {path: 'tab2', component: TabNavigationExampleChildSecond},
+                            {path: 'tab3', component: TabNavigationExampleChildThird},
+                        ]
+                    },
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.tabs }}
                 ]
             },
-            {
-                path: 'tile', component: TileDocsHeaderComponent, children: [
-                    { path: '', component: TileDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.tile } }
+            { path: 'tile', component: TileDocsHeaderComponent, children: [
+                    { path: '', component: TileDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.tile}}
                 ]
             },
-            {
-                path: 'time', component: TimeHeaderComponent, children: [
-                    { path: '', component: TimeDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.time } }
+            { path: 'time', component: TimeHeaderComponent, children: [
+                    { path: '', component: TimeDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.time}}
                 ]
             },
-            {
-                path: 'timePicker', component: TimePickerHeaderComponent, children: [
-                    { path: '', component: TimePickerDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.timePicker } }
+            { path: 'timePicker', component: TimePickerHeaderComponent, children: [
+                    { path: '', component: TimePickerDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.timePicker}}
                 ]
             },
-            {
-                path: 'tree', component: TreeHeaderComponent, children: [
-                    { path: '', component: TreeDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.tree } }
+            { path: 'tree', component: TreeHeaderComponent, children: [
+                    { path: '', component: TreeDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.tree}}
                 ]
             },
-            {
-                path: 'toggle', component: ToggleHeaderComponent, children: [
-                    { path: '', component: ToggleDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.toggle } }
+            { path: 'toggle', component: ToggleHeaderComponent, children: [
+                    { path: '', component: ToggleDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.toggle}}
                 ]
             },
-            {
-                path: 'token', component: TokenHeaderComponent, children: [
-                    { path: '', component: TokenDocsComponent },
-                    { path: 'api', component: ApiComponent, data: { content: API_FILES.token } }
+            { path: 'token', component: TokenHeaderComponent, children: [
+                    { path: '', component: TokenDocsComponent},
+                    { path: 'api', component: ApiComponent, data: {content: API_FILES.token}}
                 ]
             },
             { path: 'home', component: HomeDocsComponent },