+++ library/src/lib/modal/index.ts
@@ -1,4 +1,0 @@
-export * from './modal.module';
-export * from './modal-service/modal.service';
-export * from './modal-utils/modal-config';
-export * from './modal-utils/modal-ref';