+++ library/src/lib/side-navigation/side-navigation-group/side-navigation-group.component.ts
@@ -5,14 +5,12 @@ import { Component, ViewEncapsulation } from '@angular/core';
  * ```html
  * <fd-side-nav>
  *    <fd-side-nav-group>
- *        <fd-side-nav-title>Group Name</fd-side-nav-title>
- *          <fd-side-nav-list>    
+ *        <h1 fd-side-nav-title>Group Name</h1>
+ *          <div fd-side-nav-list>
  *             <fd-side-nav-item>
- *                <fd-side-nav-link>
- *                    <a [attr.href]="'#'">Link Item</a>
- *                </fd-side-nav-link>
+ *                <a fd-side-nav-link [attr.href]="'#'">Link Item</a>
  *             </fd-side-nav-item>
- *          </fd-side-nav-list>
+ *          </div>>
  *    </fd-side-nav-group>
  * </fd-side-nav>
  * ```